fix bunch of dumdum
